Added performance measuring to the EnergyCalculatorContainer. For a given batch size, it examines previous batch sizes whose performance have already been measured. It is a known assumption that the per-pixel calculator is faster with smaller batch sizes, and the FFT calculator is faster with larger batch sizes. If a previous batch was found to be faster with the per-pixel energy calculator, and the incoming batch size is smaller than that previous one, then the incoming batch will use the per-pixel calculator as well. On the other hand, if a previous batch was found to be faster with the FFT calculator, and the incoming batch size is larger than that previous one, then the incoming batch will use the FFT calculator.
If the incoming batch size lies somewhere in the unknown performance area, then a EnergyCalculatorMeasurer for the batch size will first measure the FFT performance, then when the next batch request comes in for that same size, it will measure the per-pixel performance. The faster calculator will then be used for all future batches of that size.